Backend/requirements.txt
flask
flask-cors
ultralytics
pillow
gunicorn
opencv-python-headless
